+++ PSGSuite/Private/EncryptionHelpers.ps1
@@ -22,6 +22,16 @@ function Get-GSDecryptedConfig {
             @{l = 'JSONServiceAccountKey'; e = {Invoke-GSDecrypt $_.JSONServiceAccountKey}}
             @{l = 'ClientSecretsPath'; e = { Invoke-GSDecrypt $_.ClientSecretsPath } }
             @{l = 'ClientSecrets'; e = { Invoke-GSDecrypt $_.ClientSecrets } }
+            @{l = 'ClientSecretOAuthScopes'; e = {
+                    $array = @()
+                    If ($_.ClientSecretOAuthScopes){
+                        foreach ($entry in $_.ClientSecretOAuthScopes){
+                            $array += Invoke-GSDecrypt $Entry
+                        }
+                    }
+                    $array
+                }
+            }
             @{l = 'AppEmail'; e = {
                     if ($_.JSONServiceAccountKey) {
                         (Invoke-GSDecrypt $_.JSONServiceAccountKey | ConvertFrom-Json).client_email